+++ crates/apollo_http_server/src/http_server.rs
@@ -27,10 +27,12 @@ use axum::handler::Handler;
 use axum::http::{HeaderMap, StatusCode};
 use axum::routing::{get, post, MethodRouter};
 use axum::{serve, Extension, Json, Router};
-use blockifier_reexecution::serde_utils::deserialize_transaction_json_to_starknet_api_tx;
 use serde::de::Error;
 use starknet_api::rpc_transaction::RpcTransaction;
-use starknet_api::serde_utils::bytes_from_hex_str;
+use starknet_api::serde_utils::{
+    bytes_from_hex_str,
+    deserialize_transaction_json_to_starknet_api_tx,
+};
 use starknet_api::transaction::fields::ValidResourceBounds;
 use tokio::net::TcpListener;
 use tokio::sync::watch::{channel, Receiver, Sender};
