+++ crates/blockifier_reexecution/src/state_reader/raw_rpc_json_test.rs
@@ -5,6 +5,7 @@ use blockifier::state::cached_state::StateMaps;
 use pretty_assertions::assert_eq;
 use rstest::{fixture, rstest};
 use starknet_api::block::{BlockInfo, BlockNumber};
+use starknet_api::serde_utils::deserialize_transaction_json_to_starknet_api_tx;
 use starknet_api::test_utils::read_json_file;
 use starknet_api::transaction::{
     DeclareTransaction,
@@ -16,7 +17,6 @@ use starknet_api::{class_hash, compiled_class_hash, contract_address, felt, nonc
 use starknet_core::types::ContractClass;
 
 use crate::compile::legacy_to_contract_class_v0;
-use crate::serde_utils::deserialize_transaction_json_to_starknet_api_tx;
 use crate::state_reader::rpc_objects::BlockHeader;
 
 #[fixture]
