+++ crates/starknet_api/src/serde_utils.rs
@@ -1,4 +1,4 @@
-//! Utilities for serialising/deserialising hexadecimal values.
+//! Utilities for serialising/deserialising values.
 #[cfg(test)]
 #[path = "serde_utils_test.rs"]
 mod serde_utils_test;
@@ -6,8 +6,15 @@ mod serde_utils_test;
 use serde::de::{Deserialize, Visitor};
 use serde::ser::{Serialize, SerializeTuple};
 use serde::Deserializer;
+use serde_json::Value;
 
 use crate::deprecated_contract_class::ContractClassAbiEntry;
+use crate::transaction::{
+    DeclareTransaction,
+    DeployAccountTransaction,
+    InvokeTransaction,
+    Transaction,
+};
 
 /// A [BytesAsHex](`crate::serde_utils::BytesAsHex`) prefixed with '0x'.
 pub type PrefixedBytesAsHex<const N: usize> = BytesAsHex<N, true>;
@@ -147,5 +154,96 @@ where
     match result {
         Ok(value) => Ok(value),
         Err(_) => Ok(None),
+    }
+}
+
+/// In old transactions, the resource bounds names are lowercase.
+/// Need to convert to uppercase for deserialization to work.
+fn upper_case_resource_bounds_names(raw_transaction: &mut Value) {
+    let resource_bounds = raw_transaction
+        .get_mut("resource_bounds")
+        .expect("tx should contain resource_bounds field")
+        .as_object_mut()
+        .expect("resource_bounds should be an object");
+
+    if let Some(l1_gas_value) = resource_bounds.remove("l1_gas") {
+        resource_bounds.insert("L1_GAS".to_string(), l1_gas_value);
+
+        let l2_gas_value = resource_bounds
+            .remove("l2_gas")
+            .expect("If tx contains l1_gas, it should contain l2_gas");
+        resource_bounds.insert("L2_GAS".to_string(), l2_gas_value);
+    }
+
+    if let Some(l1_data_gas_value) = resource_bounds.remove("l1_data_gas") {
+        resource_bounds.insert("L1_DATA_GAS".to_string(), l1_data_gas_value);
+    }
+}
+
+/// Deserializes a transaction JSON object, as returned by RPC and feeder-gateway endpoints, into
+/// a [`Transaction`]. Handles legacy formats: lowercase resource bounds names in old V3
+/// transactions, and a redundant zeroed `l1_data_gas` added by RPC v8.
+pub fn deserialize_transaction_json_to_starknet_api_tx(
+    mut raw_transaction: Value,
+) -> serde_json::Result<Transaction> {
+    let tx_type: String = serde_json::from_value(raw_transaction["type"].clone())?;
+    let tx_version: String = serde_json::from_value(raw_transaction["version"].clone())?;
+
+    // rpc_v8 fix (remove redundantly added L1DataGas)
+    let raw_resourcebounds = &raw_transaction["resource_bounds"];
+    if !raw_resourcebounds.is_null()
+        && !raw_resourcebounds["l1_data_gas"].is_null()
+        && raw_resourcebounds["l1_data_gas"]["max_amount"] == "0x0"
+        && !raw_resourcebounds["l2_gas"].is_null()
+        && raw_resourcebounds["l2_gas"]["max_amount"] == "0x0"
+    {
+        raw_transaction["resource_bounds"]
+            .as_object_mut()
+            .expect("should be map of resource bounds")
+            .remove("l1_data_gas");
+    }
+
+    match (tx_type.as_str(), tx_version.as_str()) {
+        ("INVOKE", "0x0") => {
+            Ok(Transaction::Invoke(InvokeTransaction::V0(serde_json::from_value(raw_transaction)?)))
+        }
+        ("INVOKE", "0x1") => {
+            Ok(Transaction::Invoke(InvokeTransaction::V1(serde_json::from_value(raw_transaction)?)))
+        }
+        ("INVOKE", "0x3") => {
+            // In old invoke v3 transaction, the resource bounds names are lowercase.
+            upper_case_resource_bounds_names(&mut raw_transaction);
+            Ok(Transaction::Invoke(InvokeTransaction::V3(serde_json::from_value(raw_transaction)?)))
+        }
+        ("DEPLOY_ACCOUNT", "0x1") => Ok(Transaction::DeployAccount(DeployAccountTransaction::V1(
+            serde_json::from_value(raw_transaction)?,
+        ))),
+        ("DEPLOY_ACCOUNT", "0x3") => {
+            // In old deploy account v3 transaction, the resource bounds names are lowercase.
+            upper_case_resource_bounds_names(&mut raw_transaction);
+            Ok(Transaction::DeployAccount(DeployAccountTransaction::V3(serde_json::from_value(
+                raw_transaction,
+            )?)))
+        }
+        ("DECLARE", "0x0") => Ok(Transaction::Declare(DeclareTransaction::V0(
+            serde_json::from_value(raw_transaction)?,
+        ))),
+        ("DECLARE", "0x1") => Ok(Transaction::Declare(DeclareTransaction::V1(
+            serde_json::from_value(raw_transaction)?,
+        ))),
+        ("DECLARE", "0x2") => Ok(Transaction::Declare(DeclareTransaction::V2(
+            serde_json::from_value(raw_transaction)?,
+        ))),
+        ("DECLARE", "0x3") => {
+            // In old declare v3 transaction, the resource bounds names are lowercase.
+            upper_case_resource_bounds_names(&mut raw_transaction);
+            Ok(Transaction::Declare(DeclareTransaction::V3(serde_json::from_value(
+                raw_transaction,
+            )?)))
+        }
+        ("L1_HANDLER", _) => Ok(Transaction::L1Handler(serde_json::from_value(raw_transaction)?)),
+        (tx_type, tx_version) => Err(serde::de::Error::custom(format!(
+            "unimplemented transaction type: {tx_type} version: {tx_version}"
+        ))),
     }
 }
